requirements.txt
pymongo
certifi==2019.9.11
chardet==3.0.4
configparser==4.0.2
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
kafka==1.3.5
kafka-python==1.4.7
MarkupPy==1.14
mysql-connector-python==8.0.18
odfpy==1.4.0
openpyxl==2.4.11
pbr==5.4.3
ply==3.11
protobuf==3.10.0
PyMySQL==0.9.3
PyYAML==5.1.2
records==0.5.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
tablib==0.14.0
urllib3==1.25.6
xlrd==1.2.0
xlwt==1.3.0
